requirements.txt
pytest==6.2.3
spacy==3.0.7
PyMuPDF==1.18.17
bert-extractive-summarizer==0.8.1